feat: Rename the backend-record module
